--- a/python/ray/_raylet.pyx
+++ b/python/ray/_raylet.pyx
@@ -1255,9 +1255,7 @@ cdef report_streaming_generator_output(
         # Ray Object created from an output.
         c_pair[CObjectID, shared_ptr[CRayObject]] return_obj
 
-    is_exception = isinstance(output_or_exception, Exception)
-
-    if is_exception:
+    if isinstance(output_or_exception, Exception):
         create_generator_error_object(
             output_or_exception,
             worker,
@@ -1309,8 +1307,6 @@ cdef report_streaming_generator_output(
             context.attempt_number,
             context.waiter))
 
-    return is_exception
-
 
 cdef execute_streaming_generator_sync(StreamingGeneratorExecutionContext context):
     """Execute a given generator and streaming-report the
@@ -1345,10 +1341,10 @@ cdef execute_streaming_generator_sync(StreamingGeneratorExecutionContext context
         except Exception as e:
             output_or_exception = e
 
-        done = report_streaming_generator_output(output_or_exception, context, gen_index)
+        report_streaming_generator_output(output_or_exception, context, gen_index)
         gen_index += 1
 
-        if done:
+        if isinstance(output_or_exception, Exception):
             break
 
 
@@ -1412,7 +1408,6 @@ async def execute_streaming_generator_async(
                 gen_index,
             )
         )
-
         gen_index += 1
 
         if is_exception:
@@ -1420,6 +1415,7 @@ async def execute_streaming_generator_async(
 
     # TODO elaborate
     await asyncio.gather(*futures)
+    futures.clear()
 
 
 cdef create_generator_return_obj(