--- a/python/ray/_raylet.pyx
+++ b/python/ray/_raylet.pyx
@@ -1143,9 +1143,6 @@ cdef class StreamingGeneratorExecutionContext:
         # -- Arguments that are not passed--
         # Whether or not a generator is async
         object is_async
-        # The index of the generator. E.g., if it never calls
-        # next(), it is index 0. If it calls next() once, it is 1.
-        object generator_index
         # True if `initialize` API has been called. False otherwise.
         object _is_initialized
         # -- Arguments that are passed. See the docstring for details --
@@ -1203,7 +1200,6 @@ cdef class StreamingGeneratorExecutionContext:
     ):
         cdef StreamingGeneratorExecutionContext self = (
             StreamingGeneratorExecutionContext())
-        self.generator_index = 0
         self.function_name = function_name
         self.function_descriptor = function_descriptor
         self.title = title
@@ -1232,8 +1228,10 @@ cdef class StreamingGeneratorExecutionContext:
 
 
 cdef report_streaming_generator_output(
-        output_or_exception: Union[object, Exception],
-        StreamingGeneratorExecutionContext context):
+    output_or_exception: Union[object, Exception],
+    StreamingGeneratorExecutionContext context,
+    generator_index: int64_t
+):
     """Report a given generator output to a caller.
 
     If a generator produces an exception, it should be
@@ -1256,7 +1254,6 @@ cdef report_streaming_generator_output(
     cdef:
         # Ray Object created from an output.
         c_pair[CObjectID, shared_ptr[CRayObject]] return_obj
-        int64_t generator_index = context.generator_index
 
     is_exception = isinstance(output_or_exception, Exception)
 
@@ -1274,7 +1271,7 @@ cdef report_streaming_generator_output(
             context.actor,
             context.actor_id,
             context.return_size,
-            context.generator_index,
+            generator_index,
             context.is_async,
             context.should_retry_exceptions,
             &return_obj,
@@ -1290,7 +1287,7 @@ cdef report_streaming_generator_output(
             context.caller_address,
             context.task_id,
             context.return_size,
-            context.generator_index,
+            generator_index,
             context.is_async,
             &return_obj)
 
@@ -1312,8 +1309,6 @@ cdef report_streaming_generator_output(
             context.attempt_number,
             context.waiter))
 
-    context.generator_index += 1
-
     return is_exception
 
 
@@ -1332,12 +1327,16 @@ cdef execute_streaming_generator_sync(StreamingGeneratorExecutionContext context
     Args:
         context: The context to execute streaming generator.
     """
+    cdef:
+        int64_t gen_index = 0
+
     assert context.is_initialized()
     # Generator task should only have 1 return object ref,
     # which contains None or exceptions (if system error occurs).
     assert context.return_size == 1
 
     gen = context.generator
+
     while True:
         try:
             output_or_exception = next(gen)
@@ -1346,7 +1345,9 @@ cdef execute_streaming_generator_sync(StreamingGeneratorExecutionContext context
         except Exception as e:
             output_or_exception = e
 
-        done = report_streaming_generator_output(output_or_exception, context)
+        done = report_streaming_generator_output(output_or_exception, context, gen_index)
+        gen_index += 1
+
         if done:
             break
 
@@ -1372,12 +1373,18 @@ async def execute_streaming_generator_async(
     Args:
         context: The context to execute streaming generator.
     """
+    cdef:
+        int64_t gen_index = 0
+        uint8_t is_exception = 0
+
     assert context.is_initialized()
     # Generator task should only have 1 return object ref,
     # which contains None or exceptions (if system error occurs).
     assert context.return_size == 1
 
     gen = context.generator
+
+    futures = []
     while True:
         try:
             output_or_exception = await gen.__anext__()
@@ -1388,19 +1395,31 @@ async def execute_streaming_generator_async(
             raise
         except Exception as e:
             output_or_exception = e
+            is_exception = 1
 
         loop = asyncio.get_running_loop()
         worker = ray._private.worker.global_worker
-        # Run it in a separate thread to that we can
-        # avoid blocking the event loop when serializing
-        # the output (which has nogil).
-        done = await loop.run_in_executor(
-            worker.core_worker.get_thread_pool_for_async_event_loop(),
-            report_streaming_generator_output,
-            output_or_exception,
-            context)
-        if done:
+
+        futures.append(
+            # Run it in a separate thread to that we can
+            # avoid blocking the event loop when serializing
+            # the output (which has nogil).
+            loop.run_in_executor(
+                worker.core_worker.get_thread_pool_for_async_event_loop(),
+                report_streaming_generator_output,
+                output_or_exception,
+                context,
+                gen_index,
+            )
+        )
+
+        gen_index += 1
+
+        if is_exception:
             break
+
+    # TODO elaborate
+    await asyncio.gather(*futures)
 
 
 cdef create_generator_return_obj(