@rkooo567 2024-01-25
@alexeykudinkin I remember you contributed the benchmark. Is it running daily right now?