@edoakes 2024-01-25
@rkooo567 it's not. The serve release tests are not in good shape in general. @zcin is working on improving them in the coming months. Adding these should be part of it.